Init Mobile App
I descovered I need to make custom api routes for the login and register and logut so that the mobile app also can use the next.js application for authentication. This means that I will also rewrite the login and register of the react home page on the server so that everything uses the same api endpoints. Also the web-app needs to be more structured. Will look at improving that first tmrw.